Refactor stress to enable bench and test drivers
