+++ crates/sui-benchmark/src/drivers/mod.rs
@@ -1,6 +1,5 @@
 // Copyright (c) 2022, Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod benchmark;
-pub mod drivers;
-pub mod workloads;
+pub mod bench_driver;
+pub mod driver;
